+++ config/config.js
@@ -1,0 +1,1 @@
+module.exports = require('/env/' + process.env.NODE_ENV + '.js');
